+++ examples/basic/02-iterators/main.go
@@ -6,8 +6,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/Digital-Shane/treeview"
-	"github.com/Digital-Shane/treeview/examples/shared"
+	"github.com/Digital-Shane/treeview/v2"
+	"github.com/Digital-Shane/treeview/v2/examples/shared"
 )
 
 ////////////////////////////////////////////////////////////////////