+++ examples/basic/05-workflow-visualization/main.go
@@ -5,9 +5,9 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/Digital-Shane/treeview"
-	"github.com/Digital-Shane/treeview/examples/shared"
 	"charm.land/lipgloss/v2"
+	"github.com/Digital-Shane/treeview/v2"
+	"github.com/Digital-Shane/treeview/v2/examples/shared"
 )
 
 ////////////////////////////////////////////////////////////////////