+++ examples/basic/06-builders-flat/main.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/Digital-Shane/treeview"
-	"github.com/Digital-Shane/treeview/examples/shared"
+	"github.com/Digital-Shane/treeview/v2"
+	"github.com/Digital-Shane/treeview/v2/examples/shared"
 )
 
 ////////////////////////////////////////////////////////////////////