+++ examples/intermediate/03-viewport/main.go
@@ -5,7 +5,7 @@ import (
 	"log"
 
 	tea "charm.land/bubbletea/v2"
-	"github.com/Digital-Shane/treeview"
+	"github.com/Digital-Shane/treeview/v2"
 )
 
 func main() {