+++ extensions/s3/init_test.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/ysmood/gotrace"
 
-	"github.com/Digital-Shane/treeview/extensions/s3/internal/localstack"
+	"github.com/Digital-Shane/treeview/extensions/s3/v2/internal/localstack"
 )
 
 const (