+++ extensions/s3/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.0.0] - 2026-04-22
+### Changed
+- Update major version to stay inline with treeview
+- Import path to include /v2
+
 ## [v1.0.0] - 2026-04-22
 ### Changed
 - Updates to be compatible with treeview v2.